invoice.scanner.api/requirements.txt
debugpy
langchain-community==0.3.31
langchain==0.3.27
langchain-openai==0.3.35
langchain-chroma
chromadb
openai==2.3.0
google-cloud-aiplatform==1.120.0
protobuf>=4.25.3,<5.0.0
grpcio>=1.64.0,<2.0.0
google-api-core>=2.19.1,<3.0.0
playwright==1.55.0
pycountry
flask==3.0.3
flask-smorest==0.46.2
apispec[marshmallow]==6.8.4
marshmallow<4.0.0,>=3.0.0
flask_cors
anthropic
google-generativeai
psycopg2-binary
pyodbc
jinja2>=3.0.0
python-dotenv